Rename common auth package, to reflect namespace change in -lib-auth
