+++ tracdap-libs/tracdap-lib-auth/src/main/java/org/finos/tracdap/auth/login/LoginAuthProvider.java
@@ -18,7 +18,7 @@
 package org.finos.tracdap.auth.login;
 
 import org.finos.tracdap.auth.provider.IAuthProvider;
-import org.finos.tracdap.common.auth.internal.JwtProcessor;
+import org.finos.tracdap.common.auth.JwtProcessor;
 import org.finos.tracdap.config.AuthenticationConfig;
