+++ tracdap-libs/tracdap-lib-auth/src/main/java/org/finos/tracdap/auth/login/LoginResult.java
@@ -17,7 +17,7 @@
 
 package org.finos.tracdap.auth.login;
 
-import org.finos.tracdap.common.auth.internal.UserInfo;
+import org.finos.tracdap.common.auth.UserInfo;
 import org.finos.tracdap.common.http.CommonHttpResponse;
