+++ tracdap-libs/tracdap-lib-auth/src/main/java/org/finos/tracdap/auth/login/simple/BuiltInLoginProvider.java
@@ -18,7 +18,7 @@
 package org.finos.tracdap.auth.login.simple;
 
 import org.finos.tracdap.auth.login.*;
-import org.finos.tracdap.common.auth.internal.AuthHelpers;
+import org.finos.tracdap.common.auth.AuthHelpers;
 import org.finos.tracdap.common.config.ConfigManager;
 import org.finos.tracdap.common.http.CommonHttpRequest;
 import org.finos.tracdap.common.http.Http1Headers;
