+++ tracdap-libs/tracdap-lib-auth/src/test/java/org/finos/tracdap/auth/login/GuestLoginTest.java
@@ -17,7 +17,7 @@
 
 package org.finos.tracdap.auth.login;
 
-import org.finos.tracdap.common.auth.internal.*;
+import org.finos.tracdap.common.auth.*;
 import org.finos.tracdap.config.AuthenticationConfig;
 import org.finos.tracdap.test.helpers.CloseWrapper;
 import org.finos.tracdap.test.helpers.PlatformTest;
