+++ tracdap-libs/tracdap-lib-auth/src/test/java/org/finos/tracdap/auth/login/LoginRefreshTest.java
@@ -17,7 +17,10 @@
 
 package org.finos.tracdap.auth.login;
 
-import org.finos.tracdap.common.auth.internal.*;
+import org.finos.tracdap.common.auth.JwtProcessor;
+import org.finos.tracdap.common.auth.JwtSetup;
+import org.finos.tracdap.common.auth.SessionInfo;
+import org.finos.tracdap.common.auth.UserInfo;
 import org.finos.tracdap.config.AuthenticationConfig;
 import org.finos.tracdap.test.helpers.CloseWrapper;
 import org.finos.tracdap.test.helpers.PlatformTest;
