+++ tracdap-libs/tracdap-lib-common/src/main/java/org/finos/tracdap/common/auth/AuthConstants.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.finos.tracdap.common.auth.internal;
+package org.finos.tracdap.common.auth;
 
 import io.grpc.Context;
 import io.grpc.Metadata;
