+++ tracdap-libs/tracdap-lib-common/src/main/java/org/finos/tracdap/common/auth/GrpcAuthValidator.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.finos.tracdap.common.auth.internal;
+package org.finos.tracdap.common.auth;
 
 import org.finos.tracdap.common.exception.EStartup;
 import org.finos.tracdap.config.AuthenticationConfig;
