+++ tracdap-services/tracdap-gateway/src/main/java/org/finos/tracdap/gateway/auth/Http1AuthHandler.java
@@ -22,9 +22,9 @@ import io.netty.buffer.Unpooled;
 import io.netty.util.concurrent.DefaultPromise;
 import io.netty.util.concurrent.Future;
 import io.netty.util.concurrent.Promise;
-import org.finos.tracdap.common.auth.internal.AuthHelpers;
-import org.finos.tracdap.common.auth.internal.JwtValidator;
-import org.finos.tracdap.common.auth.internal.SessionInfo;
+import org.finos.tracdap.common.auth.AuthHelpers;
+import org.finos.tracdap.common.auth.JwtValidator;
+import org.finos.tracdap.common.auth.SessionInfo;
 import org.finos.tracdap.common.exception.EUnexpected;
 import org.finos.tracdap.common.http.Http1Headers;
 import org.finos.tracdap.common.netty.ConnectionId;
