+++ tracdap-services/tracdap-svc-auth/src/main/java/org/finos/tracdap/svc/auth/ProviderLookup.java
@@ -20,7 +20,7 @@ package org.finos.tracdap.svc.auth;
 import org.finos.tracdap.auth.login.ILoginProvider;
 import org.finos.tracdap.auth.login.LoginAuthProvider;
 import org.finos.tracdap.auth.provider.IAuthProvider;
-import org.finos.tracdap.common.auth.internal.JwtSetup;
+import org.finos.tracdap.common.auth.JwtSetup;
 import org.finos.tracdap.common.config.ConfigManager;
 import org.finos.tracdap.common.exception.EStartup;
 import org.finos.tracdap.common.exception.EUnexpected;
