+++ tracdap-services/tracdap-svc-auth/src/test/java/org/finos/tracdap/svc/auth/ProviderSwitchingTest.java
@@ -17,8 +17,8 @@
 
 package org.finos.tracdap.svc.auth;
 
-import org.finos.tracdap.common.auth.internal.JwtSetup;
-import org.finos.tracdap.common.auth.internal.JwtValidator;
+import org.finos.tracdap.common.auth.JwtSetup;
+import org.finos.tracdap.common.auth.JwtValidator;
 import org.finos.tracdap.test.helpers.CloseWrapper;
 import org.finos.tracdap.test.helpers.PlatformTest;
 import org.junit.jupiter.api.Assertions;
