+++ tracdap-services/tracdap-svc-data/src/main/java/org/finos/tracdap/svc/data/TracDataService.java
@@ -19,9 +19,9 @@ package org.finos.tracdap.svc.data;
 
 import org.finos.tracdap.api.internal.TrustedMetadataApiGrpc;
 import org.finos.tracdap.common.netty.*;
-import org.finos.tracdap.common.auth.internal.InternalAuthProvider;
-import org.finos.tracdap.common.auth.internal.JwtSetup;
-import org.finos.tracdap.common.auth.internal.GrpcAuthValidator;
+import org.finos.tracdap.common.auth.InternalAuthProvider;
+import org.finos.tracdap.common.auth.JwtSetup;
+import org.finos.tracdap.common.auth.GrpcAuthValidator;
 import org.finos.tracdap.common.config.ConfigKeys;
 import org.finos.tracdap.common.grpc.*;
 import org.finos.tracdap.common.codec.CodecManager;
