+++ tracdap-services/tracdap-svc-data/src/main/java/org/finos/tracdap/svc/data/api/TracDataApi.java
@@ -20,7 +20,7 @@ package org.finos.tracdap.svc.data.api;
 import org.finos.tracdap.api.*;
 import org.finos.tracdap.api.Data;
 import org.finos.tracdap.common.netty.EventLoopResolver;
-import org.finos.tracdap.common.auth.internal.GrpcAuthHelpers;
+import org.finos.tracdap.common.auth.GrpcAuthHelpers;
 import org.finos.tracdap.common.data.DataContext;
 import org.finos.tracdap.common.data.IDataContext;
 import org.finos.tracdap.common.data.pipeline.GrpcDownloadSink;
