+++ tracdap-services/tracdap-svc-data/src/main/java/org/finos/tracdap/svc/data/service/RequestState.java
@@ -18,7 +18,7 @@
 package org.finos.tracdap.svc.data.service;
 
 import io.grpc.CallCredentials;
-import org.finos.tracdap.common.auth.internal.UserInfo;
+import org.finos.tracdap.common.auth.UserInfo;
 import org.finos.tracdap.metadata.*;
 
 import java.time.Instant;
