+++ tracdap-services/tracdap-svc-orch/src/main/java/org/finos/tracdap/svc/orch/service/JobProcessor.java
@@ -22,8 +22,8 @@ import org.finos.tracdap.api.JobStatus;
 import org.finos.tracdap.api.MetadataWriteRequest;
 import org.finos.tracdap.api.internal.RuntimeJobStatus;
 import org.finos.tracdap.api.internal.TrustedMetadataApiGrpc.TrustedMetadataApiBlockingStub;
-import org.finos.tracdap.common.auth.internal.GrpcAuthHelpers;
-import org.finos.tracdap.common.auth.internal.InternalAuthProvider;
+import org.finos.tracdap.common.auth.GrpcAuthHelpers;
+import org.finos.tracdap.common.auth.InternalAuthProvider;
 import org.finos.tracdap.common.cache.CacheEntry;
 import org.finos.tracdap.common.exception.*;
 import org.finos.tracdap.common.exec.*;
