+++ tracdap-services/tracdap-svc-orch/src/main/java/org/finos/tracdap/svc/orch/service/JobState.java
@@ -20,8 +20,8 @@ package org.finos.tracdap.svc.orch.service;
 import org.finos.tracdap.api.JobRequest;
 import org.finos.tracdap.api.internal.RuntimeJobResult;
 import org.finos.tracdap.api.internal.RuntimeJobStatus;
-import org.finos.tracdap.common.auth.internal.InternalCallCredentials;
-import org.finos.tracdap.common.auth.internal.UserInfo;
+import org.finos.tracdap.common.auth.InternalCallCredentials;
+import org.finos.tracdap.common.auth.UserInfo;
 import org.finos.tracdap.common.exception.EUnexpected;
 import org.finos.tracdap.config.JobConfig;
 import org.finos.tracdap.config.RuntimeConfig;
